HSEARCH-1188 Creating infrastructure for running tests from "orm" module against ES backend
