+++ elasticsearch/src/test/java/org/hibernate/search/backend/elasticsearch/test/ElasticSearchIT.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.test.backend.elasticsearch;
+package org.hibernate.search.backend.elasticsearch.test;
 
 import java.math.BigInteger;
 import java.util.Arrays;
@@ -23,6 +23,7 @@ import org.hibernate.search.FullTextSession;
 import org.hibernate.search.Search;
 import org.hibernate.search.backend.elasticsearch.ElasticSearchQueries;
 import org.hibernate.search.backend.elasticsearch.ProjectionConstants;
+import org.hibernate.search.backend.elasticsearch.testutil.JsonHelper;
 import org.hibernate.search.query.engine.spi.QueryDescriptor;
 import org.hibernate.search.test.SearchTestBase;
 import org.hibernate.search.test.embedded.Address;
@@ -32,7 +33,6 @@ import org.hibernate.search.test.embedded.Person;
 import org.hibernate.search.test.embedded.State;
 import org.hibernate.search.test.embedded.StateCandidate;
 import org.hibernate.search.test.embedded.Tower;
-import org.hibernate.search.testutil.backend.elasticsearch.JsonHelper;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
