+++ elasticsearch/src/test/java/org/hibernate/search/backend/elasticsearch/test/ResearchPaper.java
@@ -4,22 +4,23 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.test.backend.elasticsearch;
+package org.hibernate.search.backend.elasticsearch.test;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 
+import org.hibernate.search.annotations.Boost;
 import org.hibernate.search.annotations.DocumentId;
 import org.hibernate.search.annotations.Field;
 import org.hibernate.search.annotations.Indexed;
 
 /**
- * @author Gunnar Morling
+ * @author Richard Hallier
  */
 @Entity
-@Indexed(index = "studentwork")
-public class MasterThesis {
+@Indexed
+public class ResearchPaper {
 
 	@Id
 	@GeneratedValue
@@ -27,17 +28,26 @@ public class MasterThesis {
 	private Long id;
 
 	@Field
+	@Boost(2)
 	private String title;
+
+	@Field(name = "abstract")
+	private String summary;
 
 	@Field
 	private String text;
 
-	MasterThesis() {
+	@Field
+	private int wordCount;
+
+	ResearchPaper() {
 	}
 
-	public MasterThesis(String title, String text) {
+	public ResearchPaper(String title, String summary, String text, int wordCount) {
+		this.summary = summary;
 		this.text = text;
 		this.title = title;
+		this.wordCount = wordCount;
 	}
 
 	public Long getId() {
@@ -56,11 +66,27 @@ public class MasterThesis {
 		this.title = title;
 	}
 
+	public String getSummary() {
+		return summary;
+	}
+
+	public void setSummary(String summary) {
+		this.summary = summary;
+	}
+
 	public String getText() {
 		return text;
 	}
 
 	public void setText(String text) {
 		this.text = text;
+	}
+
+	public int getWordCount() {
+		return wordCount;
+	}
+
+	public void setWordCount(int wordCount) {
+		this.wordCount = wordCount;
 	}
 }
