+++ pom.xml
@@ -646,6 +646,11 @@
                 <artifactId>xml-apis</artifactId>
                 <version>1.3.03</version>
             </dependency>
+            <dependency>
+                <groupId>io.takari.junit</groupId>
+                <artifactId>takari-cpsuite</artifactId>
+                <version>1.2.7</version>
+            </dependency>
         </dependencies>
     </dependencyManagement>
